Refactor MVC after OSMH review
