+++ ImageCarousel/Randomizer/images.txt
@@ -1,4 +1,2 @@
-2.jpg
-3.jpg
-pp.jpg
+4.jpg
