--- ImageCarousel/images.txt
@@ -1,3 +1,0 @@
-2.jpg
-pp.jpg
-
